Dark theme for the markdown preview pane
Added dark theme for the preview markdown pane. This is enabled by default when the theme name contains the word "dark", and automatically updates to either dark or light mode when a new theme is applied.
In the future, themes will have dark/light mode tags which will be used instead.